requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.10.0
uvicorn==0.29.0
